Extract a shared ContainerConfigForm from Run/Edit
The Run and Edit container views duplicated ~1000 lines: the ConfigTab enum,
tab picker, five tab content views, empty-state, row add/remove, and the row
components. Extract it all into ContainerConfigForm, parameterized by a mode
(.run/.edit) that drives the few real differences — editable+validated name
and DNS/network pickers on run, locked name + recreate warning + ports note
on edit. Run/Edit shrink to their chrome (header/footer, plus edit's warning)
and their action (runContainer / recreateContainer).

Name validation becomes a pure static function
(ContainerConfigForm.validationError(for:existing:)) with unit tests.

RunContainer 630->114, EditContainer 507->186, +523 shared. Build green;
97->102 tests pass.
